First "working" version of the complete script.
Lots of tests to make. It will also be useful to restructure the files (no need for many folders, one for each day is enough). Or I could also try uploading the files directly to the server - more difficult and more risky; will add this once the rates are well tested. The one-folder-for-all will make things easier for this, too.